% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/InferenceEngine-methods.R
\docType{methods}
\name{add.observations<-}
\alias{add.observations<-}
\alias{add.observations<-,InferenceEngine-method}
\title{add further evidence to an existing list of observations of an \code{\link{InferenceEngine}}.}
\usage{
add.observations(x) <- value

\S4method{add.observations}{InferenceEngine}(x) <- value
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}

\item{value}{the list of observations of the \code{\link{InferenceEngine}}.}
}
\description{
Add a list of observations to an InferenceEngine that already has observations,
using a list composed by the two following vectors:
\itemize{
\item{\code{observed.vars}}{vector of observed variables;}
\item{\code{observed.vals}}{vector of values observed for the variables in \code{observed.vars} in the corresponding position.}
}
}
\details{
In case of multiple observations of the same variable, the last observation is the one used, as the most recent.
}
\seealso{
\code{\link{observations<-}}
}
