% Generated by roxygen2 (4.0.1): do not edit by hand
\name{asia}
\alias{asia}
\title{load \code{Asia} dataset.}
\usage{
asia()
}
\value{
a BNDataset containing the \code{Child} dataset.
}
\description{
Wrapper for a loader for the \code{Asia} dataset, with only raw data.
}
\details{
The dataset has 10000 items, no missing data, so no imputation needs to be performed.
}
\examples{
dataset <- asia()
print(dataset)
}
\seealso{
\code{\link{asia_10000}}
}

