% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{asia_10000}
\alias{asia_10000}
\title{\code{Asia} dataset.}
\format{
a \code{\link{BNDataset}} with raw data slow filled.
}
\description{
The \code{Asia} dataset contains 10000 complete (no missing data, no latent variables) randomly generated items of the \code{Asia} Bayesian Network.
No imputation needs to be performed, so only raw data is present.
}
\details{
The data the BNDataset object is built from is located in files \code{pkg_folder/extdata/asia_10000.header} and \code{pkg_folder/extdata/asia_10000.data}.
}
\references{
S. Lauritzen, D. Spiegelhalter. Local Computation with Probabilities on Graphical Structures and their Application to Expert Systems
(with discussion). Journal of the Royal Statistical Society: Series B (Statistical Methodology), 50(2):157-224, 1988.
}
\seealso{
\code{\link{asia}}
}
