% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-methods.R
\name{asia_2_layers}
\alias{asia_2_layers}
\title{load a two-layers dataset derived from the \code{Asia} dataset.}
\usage{
asia_2_layers()
}
\value{
a BNDataset containing the \code{Child} dataset.
}
\description{
Wrapper for a loader for a 2-layers dataset derived from the \code{Asia} dataset, with only raw data.
}
\details{
The dataset has 100 items, no missing data, so no imputation needs to be performed.
}
\examples{
dataset <- asia_2_layers()
print(dataset)

}
\seealso{
\code{\link{asia_10000}}
}
