% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{belief.propagation}
\alias{belief.propagation}
\alias{belief.propagation,InferenceEngine}
\alias{belief.propagation,InferenceEngine-method}
\title{perform belief propagation.}
\usage{
belief.propagation(ie, net = NULL, observed.vars = NULL,
  observed.vals = NULL, return.potentials = FALSE, ...)

\S4method{belief.propagation}{InferenceEngine}(ie, net = NULL,
  observed.vars = NULL, observed.vals = NULL, return.potentials = FALSE,
  ...)
}
\arguments{
\item{ie}{an \code{\link{InferenceEngine}} object.}

\item{net}{a \code{\link{BN}} object.}

\item{observed.vars}{list of observed variables.}

\item{observed.vals}{values taken by variables listed in \code{observed.vars}.}

\item{return.potentials}{if TRUE only the potentials are returned, instead of the default \code{\link{BN}}.}

\item{...}{potential further arguments of methods.}
}
\value{
updated \code{\link{InferenceEngine}} object.
}
\description{
Perform belief propagation for the network of an InferenceEngine, given a set of observations when present.
In the current version of \code{bnstruct}, belief propagation can be computed only over a junction tree.
}
\examples{
\dontrun{
dataset <- BNDataset()
dataset <- read.dataset(dataset, "file.header", "file.data")
bn <- BN(dataset)
ie <- InferenceEngine(bn)
ie <- belief.propagation(ie)

observations(ie) <- list("observed.vars"=("A","G","X"), "observed.vals"=c(1,2,1))
belief.propagation(ie)
}
}

