% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{BNParams-class}
\alias{BNParams}
\alias{BNParams,BNParams-class}
\alias{BNParams-class}
\alias{initialize,BNParams-method}
\title{BNParams class}
\usage{
\S4method{initialize}{BN}(.Object, ...)

BNParams(args, ...)
}
\arguments{
\item{.Object}{a BNParams}

\item{...}{potential further arguments of methods.}

\item{args}{the list of command line parameters passed to the R script.}
}
\value{
BNParams object.
}
\description{
Contains all of the parameters available for the algorithms contained in the package.

Takes the list of command line parameters passed in input, and processes it.
Does not check for their correctness (it's assumed they are correct).
}
\section{Slots}{

\describe{
\item{\code{CPX_PARAM_SCRIND}}{= "numeric",}

\item{\code{CPX_PARAM_PRELINEAR}}{= "numeric",}

\item{\code{CPX_PARAM_PREIND}}{= "numeric",}

\item{\code{CPX_PARAM_MIPCBREDLP}}{= "numeric",}

\item{\code{CPX_PARAM_MIPSEARCH}}{= "numeric",}

\item{\code{cplex_presolve_algo}}{= "numeric"}
}}
\examples{
\dontrun{
params <- BNParams()
}
}

