% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io-methods.R
\name{bnstruct.reset.log}
\alias{bnstruct.reset.log}
\title{reset internal logging methods}
\usage{
bnstruct.reset.log()
}
\description{
Reset internal logging methods, e.g. in case of user interruption.
}

