% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{boots<-}
\alias{boots<-}
\alias{boots<-,BNDataset-method}
\title{set list of bootstrap samples of a \code{\link{BNDataset}}.}
\usage{
boots(x) <- value
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}

\item{value}{the list of bootstrap samples.}
}
\description{
Add to a dataset a list of samples from raw data computed using bootstrap.
}

