% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{boots}
\alias{boots}
\alias{boots,BNDataset}
\alias{boots,BNDataset-method}
\title{get list of bootstrap samples of a \code{\link{BNDataset}}.}
\usage{
boots(x)

\S4method{boots}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}
}
\value{
the list of bootstrap samples.
}
\description{
Return the list of samples computed from raw data of a dataset.
}
\seealso{
\code{\link{has.boots}}, \code{\link{has.imp.boots}}, \code{\link{imp.boots}}
}

