% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\docType{methods}
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap,BNDataset}
\alias{bootstrap,BNDataset-method}
\title{Perform bootstrap.}
\usage{
bootstrap(object, num.boots = params@num.boots, seed = params@seed,
  imputation = FALSE, k.impute = params@k.impute, params)

\S4method{bootstrap}{BNDataset}(object, num.boots = params@num.boots,
  seed = params@seed, imputation = FALSE, k.impute = params@k.impute,
  params)
}
\arguments{
\item{object}{the \code{\link{BNDataset}} object.}

\item{num.boots}{number of sampled datasets for bootstrap.}

\item{seed}{random seed.}

\item{imputation}{\code{TRUE} if imputation has to be performed. Default is \code{FALSE}.}

\item{k.impute}{number of neighbours to be used; for discrete variables we use mode, for continuous variables
the median value is instead taken (useful only if imputation == TRUE).}

\item{params}{parameters.}
}
\description{
Create a list of \code{num.boots} samples of the original dataset.
}
\examples{
\dontrun{
dataset <- BNDataset("file.data", "file.header")
dataset <- bootstrap(dataset, num.boots = 1000)
}
}

