% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-methods.R
\name{child}
\alias{child}
\title{load \code{Child} dataset.}
\usage{
child()
}
\value{
a BNDataset containing the \code{Child} dataset.
}
\description{
Wrapper for a loader for the \code{Child} raw dataset; also perform imputation.
}
\details{
The dataset has 5000 items, with random missing values (no latent variables). BNDataset object contains the raw dataset and imputed dataset, with \code{k=10}
(see \code{\link{impute}} for related explanation).
}
\examples{
dataset <- child()
print(dataset)

}
\seealso{
\code{\link{child_NA_5000}}
}
