% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{child_NA_5000}
\alias{child_NA_5000}
\title{\code{Child} dataset.}
\format{
a \code{\link{BNDataset}} with a raw and imputed data slow filled with 5000 items.
}
\description{
The \code{Child} dataset contains 5000 randomly generated items with missing data (no latent variables) of the \code{Child} Bayesian Network.
Imputation is performed, so both raw and imputed data is present.
}
\details{
The data the BNDataset object is built from is located in files \code{pkg_folder/extdata/extdata/Child_data_na_5000.header} and \code{pkg_folder/extdata/extdata/Child_data_na_5000.data}.
}
\references{
D. J. Spiegelhalter, R. G. Cowell (1992). Learning in probabilistic expert systems. In Bayesian Statistics 4
(J. M. Bernardo, J. 0. Berger, A. P. Dawid and A. F. M. Smith, eds.) 447-466. Clarendon Press, Oxford.
}
\seealso{
\code{\link{child}}
}
