% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R
\docType{methods}
\name{dag<-}
\alias{dag<-}
\alias{dag<-,BN-method}
\title{set adjacency matrix of an object.}
\usage{
dag(x) <- value

\S4method{dag}{BN}(x) <- value
}
\arguments{
\item{x}{an object.}

\item{value}{matrix containing the adjacency matrix of the directed acyclic graph representing
the structure of the object.}
}
\description{
Set the adjacency matrix of the directed acyclic graph representing the structure of a network.
}
