% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\docType{methods}
\name{data.file<-}
\alias{data.file<-}
\alias{data.file<-,BNDataset-method}
\title{set data file of a \code{\link{BNDataset}}.}
\usage{
data.file(x) <- value

\S4method{data.file}{BNDataset}(x) <- value
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}

\item{value}{data filename.}
}
\description{
Set the data filename of a dataset (with the path to its position, as given by the user).
The data filename may contain a header in the first row, containing the list of names of the variables,
in the same order as in the header file.
After the header, if present, the file contains a data.frame with the observations, one item per row.
}
\seealso{
\code{\link{header.file<-}}
}
