% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\name{data.file}
\alias{data.file}
\alias{data.file,BNDataset-method}
\alias{data.file,BNDataset}
\title{get data file of a \code{\link{BNDataset}}.}
\usage{
data.file(x)

\S4method{data.file}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}
}
\value{
data filename of the dataset.
}
\description{
Return the data filename of a dataset (with the path to its position, as given by the user).
The data filename may contain a header in the first row, containing the list of names of the variables,
in the same order as in the header file.
After the header, if present, the file contains a data.frame with the observations, one item per row.
}
\seealso{
\code{\link{data.file}}
}
