% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{discreteness<-}
\alias{discreteness<-}
\alias{discreteness<-,BN-method}
\alias{discreteness<-,BNDataset-method}
\title{set status (discrete or continuous) of the variables of an object.}
\usage{
discreteness(x) <- value
}
\arguments{
\item{x}{an object.}

\item{value}{a vector of elements in \{\code{c},\code{d}\} for continuous and discrete variables (respectively).}
}
\description{
Set the list of variable status for the variables in a network or a dataset.
}

