% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN-methods.R
\name{edge.dir.wpdag}
\alias{edge.dir.wpdag}
\title{counts the edges in a WPDAG with their directionality}
\usage{
edge.dir.wpdag(x, use.node.names = TRUE)
}
\arguments{
\item{x}{the \code{BN}}

\item{use.node.names}{use node names rather than number (\code{TRUE} by default).}
}
\value{
a matrix containing the node pairs with the count of the edges
        between them in the \code{WPDAG}.
}
\description{
Given a \code{BN} with a \code{WPDAG}, it counts the edges, with
their directionality.
}
