% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{get.data}
\alias{get.data}
\alias{get.data,BNDataset}
\alias{get.data,BNDataset-method}
\title{get data of a BNDataset.}
\usage{
get.data(x)

\S4method{get.data}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}
}
\description{
Return data contained in a \code{\link{BNDataset}} object, if any.
Preference is given to imputed data, if available, because the imputed dataset
is (supposed to be), in general, more useful. To obtain specifically raw or imputed data,
one must revert to \code{\link{get.raw.data}()} and \code{\link{get.imputed.data}()}, respectively.
}
\examples{
\dontrun{
x <- BNDataset()
x <- read.dataset(x, "file.header", "file.data")
get.data(x) # returns raw dataset, the only one present in dataset

x <- impute(x)
get.data(x) # returns imputed dataset, since it is present now
}
}
\seealso{
\code{\link{has.data}}, \code{\link{has.raw.data}}, \code{\link{has.imputed.data}}, \code{\link{get.raw.data}}, \code{\link{get.imputed.data}}
}

