% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get.imputed.data}
\alias{get.imputed.data}
\alias{get.imputed.data,BNDataset}
\alias{get.imputed.data,BNDataset-method}
\title{get imputed data of a BNDataset.}
\usage{
get.imputed.data(x)

\S4method{get.imputed.data}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}
}
\description{
Return imputed data contained in a \code{\link{BNDataset}} object, if any.
}
\seealso{
\code{\link{has.data}}, \code{\link{has.raw.data}}, \code{\link{has.imputed.data}}, \code{\link{get.data}}, \code{\link{get.raw.data}}
}

