% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{has.data}
\alias{has.data}
\alias{has.data,BNDataset}
\alias{has.data,BNDataset-method}
\title{check if a BNDataset contains any data.}
\usage{
has.data(x)

\S4method{has.data}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}
}
\description{
Check whether a \code{\link{BNDataset}} object actually contains raw or imputed data.
}
\examples{
\dontrun{
x <- BNDataset()
has.data(x) # FALSE

x <- read.dataset(x, "file.header", "file.data")
has.data(x) # TRUE
}
}
\seealso{
\code{\link{has.raw.data}}, \code{\link{has.imputed.data}}, \code{\link{get.data}}, \code{\link{get.raw.data}}, \code{\link{get.imputed.data}}
}

