% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{has.imp.boots}
\alias{has.imp.boots}
\alias{has.imp.boots,BNDataset}
\alias{has.imp.boots,BNDataset-method}
\title{check whether a \code{\link{BNDataset}} has bootstrap samples from imputed data or not.}
\usage{
has.imp.boots(x)

\S4method{has.imp.boots}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}
}
\value{
\code{TRUE} if dataset has bootstrap samples from imputed data.
}
\description{
Return \code{TRUE} if the given dataset contains samples for bootstrap from inputed dataset, \code{FALSE} otherwise.
}
\seealso{
\code{\link{has.boots}}, \code{\link{boots}}, \code{\link{imp.boots}}
}

