% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\name{has.imputed.boots}
\alias{has.imputed.boots}
\alias{has.imputed.boots,BNDataset-method}
\alias{has.imputed.boots,BNDataset}
\title{check whether a \code{\link{BNDataset}} has bootstrap samples from imputed data or not.}
\usage{
has.imputed.boots(x)

\S4method{has.imputed.boots}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}
}
\value{
\code{TRUE} if dataset has bootstrap samples from imputed data.
}
\description{
Return \code{TRUE} if the given dataset contains samples for bootstrap from inputed dataset, \code{FALSE} otherwise.
}
\seealso{
\code{\link{has.boots}}, \code{\link{boots}}, \code{\link{imp.boots}}
}
