% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\docType{methods}
\name{has.raw.data}
\alias{has.raw.data}
\alias{has.raw.data,BNDataset}
\alias{has.raw.data,BNDataset-method}
\title{check if a BNDataset contains raw data.}
\usage{
has.raw.data(x)

has.raw.data(x)

\S4method{has.raw.data}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}

\item{x}{a \code{\link{BNDataset}} object.}
}
\value{
\code{TRUE} if dataset has raw data.
}
\description{
Check whether a \code{\link{BNDataset}} object actually contains raw data.

Return \code{TRUE} if the given dataset contains raw data, \code{FALSE} otherwise.
}
\examples{
\dontrun{
x <- BNDataset()
has.raw.data(x) # FALSE

x <- read.dataset(x, "file.header", "file.data")
has.raw.data(x) # TRUE, since read.dataset() actually reads raw data.
}
}
\seealso{
\code{\link{has.imputed.data}}, \code{\link{raw.data}}, \code{\link{imputed.data}}

\code{\link{has.imputed.data}}, \code{\link{raw.data}}, \code{\link{imputed.data}}
}

