% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\docType{methods}
\name{header.file<-}
\alias{header.file<-}
\alias{header.file<-,BNDataset-method}
\title{set header file of a \code{\link{BNDataset}}.}
\usage{
header.file(x) <- value

\S4method{header.file}{BNDataset}(x) <- value
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}

\item{value}{header filename.}
}
\description{
Set the header filename of a dataset (with the path to its position, as given by the user).
The header file has to contain three rows:
\enumerate{
\item list of names of the variables, in the same order as in the data file;
\item list of cardinalities of the variables, if discrete, or levels for quantization if continuous;
\item list of status of the variables: \code{c} for continuous variables, \code{d} for discrete ones.
}
Further rows are ignored.
}
\seealso{
\code{\link{data.file<-}}
}
