% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\name{header.file}
\alias{header.file}
\alias{header.file,BNDataset-method}
\alias{header.file,BNDataset}
\title{get header file of a \code{\link{BNDataset}}.}
\usage{
header.file(x)

\S4method{header.file}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}
}
\value{
header filename of the dataset.
}
\description{
Return the header filename of a dataset (with the path to its position, as given by the user),
present if the dataset has been read from a file and not manually inserted.
The header file contains three rows:
\enumerate{
\item list of names of the variables, in the same order as in the data file;
\item list of cardinalities of the variables, if discrete, or levels for quantization if continuous;
\item list of status of the variables: \code{c} for continuous variables, \code{d} for discrete ones.
}
}
\seealso{
\code{\link{data.file}}
}
