% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\docType{methods}
\name{imp.boots<-}
\alias{imp.boots<-}
\alias{imp.boots<-,BNDataset-method}
\title{set list of bootstrap samples from imputed data of a \code{\link{BNDataset}}.}
\usage{
imp.boots(x) <- value

\S4method{imp.boots}{BNDataset}(x) <- value
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}

\item{value}{the list of bootstrap samples from imputed data.}
}
\description{
Add to a dataset a list of samples from imputed data computed using bootstrap.
}
