% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{imputed.data<-}
\alias{imputed.data<-}
\alias{imputed.data<-,BNDataset-method}
\title{add imputed data.}
\usage{
imputed.data(x) <- value

\S4method{imputed.data}{BNDataset}(x) <- value
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}

\item{value}{a matrix of integers containing a dataset.}
}
\description{
Insert imputed data in a \code{\link{BNDataset}} object.
}
\details{
Users are encouraged to not use this method whenever possible, in favour of \code{\link{read.dataset}} with flag \code{imputation = TRUE}.
}
\seealso{
\code{\link{has.data}}, \code{\link{has.imputed.data}}, \code{\link{get.data}}, \code{\link{read.dataset}}
}

