% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\name{imputed.data}
\alias{imputed.data}
\alias{imputed.data,BNDataset-method}
\alias{imputed.data,BNDataset}
\title{get imputed data of a BNDataset.}
\usage{
imputed.data(x)

\S4method{imputed.data}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}
}
\description{
Return imputed data contained in a \code{\link{BNDataset}} object, if any.
}
\seealso{
\code{\link{has.raw.data}}, \code{\link{has.imputed.data}}, \code{\link{raw.data}}
}
