% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/InferenceEngine-methods.R
\name{interventions}
\alias{interventions}
\alias{interventions,InferenceEngine-method}
\alias{interventions,InferenceEngine}
\title{get the list of interventions of an \code{\link{InferenceEngine}}.}
\usage{
interventions(x)

\S4method{interventions}{InferenceEngine}(x)
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}
}
\value{
the list of interventions of the \code{\link{InferenceEngine}}.
}
\description{
Return the list of interventions added to an InferenceEngine.
}
\details{
Output is a list in the following format:
\itemize{
\item{\code{intervention.vars}}{vector of manipulated variables;}
\item{\code{intervention.vals}}{vector of values for the variables in \code{observed.vars} in the corresponding position.}
}
}
