% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/InferenceEngine-methods.R
\docType{methods}
\name{jpts<-}
\alias{jpts<-}
\alias{jpts<-,InferenceEngine-method}
\title{set the list of joint probability tables compiled by an \code{\link{InferenceEngine}}.}
\usage{
jpts(x) <- value

\S4method{jpts}{InferenceEngine}(x) <- value
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}

\item{value}{the list of joint probability tables compiled by the \code{\link{InferenceEngine}}.}
}
\description{
Add a list of joint probability tables for the cliques of the junction tree.
}
\details{
Each joint probability table is represented as a multidimensional array. 
To retrieve single dimensions (e.g. to compute marginals), users should provide dimension names.
}
