% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/InferenceEngine-methods.R
\name{jpts}
\alias{jpts}
\alias{jpts,InferenceEngine-method}
\alias{jpts,InferenceEngine}
\title{get the list of joint probability tables compiled by an \code{\link{InferenceEngine}}.}
\usage{
jpts(x)

\S4method{jpts}{InferenceEngine}(x)
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}
}
\value{
the list of joint probability tables compiled by the \code{\link{InferenceEngine}}.
}
\description{
Return the list of joint probability tables for the cliques of the junction tree 
obtained after belief propagation has been performed.
}
\details{
Each joint probability table is represented as a multidimensional array. 
To retrieve single dimensions (e.g. to compute marginals), users should not rely on dimension numbers,
but should instead select the dimensions using their names.
}
