% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{jt.cliques<-}
\alias{jt.cliques<-}
\alias{jt.cliques<-,InferenceEngine-method}
\title{set the list of cliques of the junction tree of an \code{\link{InferenceEngine}}.}
\usage{
jt.cliques(x) <- value
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}

\item{value}{the list of cliques of the junction tree contained in the \code{\link{InferenceEngine}}.}
}
\description{
Add to the InferenceEngine a list containing the cliques of variables composing the nodes of the junction tree.
}

