% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{jt.cliques}
\alias{jt.cliques}
\alias{jt.cliques,InferenceEngine}
\alias{jt.cliques,InferenceEngine-method}
\title{get the list of cliques of the junction tree of an \code{\link{InferenceEngine}}.}
\usage{
jt.cliques(x)

\S4method{jt.cliques}{InferenceEngine}(x)
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}
}
\value{
the list of cliques of the junction tree contained in the \code{\link{InferenceEngine}}.
}
\description{
Return the list of cliques containing the variables associated to each node of a junction tree.
}

