% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/InferenceEngine-methods.R
\name{junction.tree}
\alias{junction.tree}
\alias{junction.tree,InferenceEngine-method}
\alias{junction.tree,InferenceEngine}
\title{get the junction tree of an \code{\link{InferenceEngine}}.}
\usage{
junction.tree(x)

\S4method{junction.tree}{InferenceEngine}(x)
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}
}
\value{
the junction tree contained in the \code{\link{InferenceEngine}}.
}
\description{
Return the adjacency matrix representing the junction tree computed for a network.
}
\details{
Rows and columns are named after the (variables in the) cliques that each node of the junction tree represent.
}
\seealso{
\code{\link{build.junction.tree}}
}
