% Generated by roxygen2 (4.0.1): do not edit by hand
\name{knn.impute}
\alias{knn.impute}
\title{Perform imputation of a data frame using k-NN.}
\usage{
knn.impute(data, k, cat.var)
}
\arguments{
\item{data}{a data frame}

\item{k}{number of neighbours to be used; for discrete variables we use mode, for continuous variables the median value is instead taken.}

\item{cat.var}{vector containing the cardinality of the variables, if discrete,
       or the number of levels the variables are going to be quantized into, if continuous.}
}
\value{
imputed data frame.
}
\description{
Perform imputation of missing data in a data frame using the k-Nearest Neighbour algorithm.
For discrete variables we use the mode, for continuous variables the median value is instead taken.
}

