% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{layering}
\alias{layering}
\alias{layering,BN}
\alias{layering,BN-method}
\alias{layering,InferenceEngine}
\alias{layering,InferenceEngine-method}
\title{return the layering of the nodes.}
\usage{
layering(x, ...)

\S4method{layering}{BN}(x, ...)

\S4method{layering}{InferenceEngine}(x, updated.bn = TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{BN}} or \code{\link{InferenceEngine}} object.}

\item{updated.bn}{\code{TRUE} if \code{x} is an InferenceEngine and the updated network is chosen (kept only for compatibility with other methods).}

\item{...}{potential further arguments for methods.}
}
\value{
a vector containing layers the nodes can be divided into.
}
\description{
Compute the topological ordering of the nodes of a network, in order to divide the network in layers.
}
\examples{
\dontrun{
dataset <- BNDataset(name="MyDataset")
dataset <- read.dataset(dataset, "file.header", "file.data")
x <- BN(dataset)
layering(x)
eng <- InferenceEngine(x)
layering(x, updated.bn=TRUE)
}
}

