% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{name<-}
\alias{name<-}
\alias{name<-,BN-method}
\alias{name<-,BNDataset-method}
\title{set name of an object.}
\usage{
name(x) <- value
}
\arguments{
\item{x}{an object.}

\item{value}{the new name of the object.}
}
\description{
Set the \code{name} slot of an object of type \code{\link{BN}} or \code{\link{BNDataset}}.
}

