% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{node.sizes<-}
\alias{node.sizes<-}
\alias{node.sizes<-,BN-method}
\alias{node.sizes<-,BNDataset-method}
\title{set the size of variables of an object.}
\usage{
node.sizes(x) <- value
}
\arguments{
\item{x}{an object.}

\item{value}{vector contaning the size of each variable of the object.}
}
\description{
Set the size of the variables of a BN or BNDataset object. It represents the actual cardinality
of discrete variables, and the cardinality of the discretized variable for continuous variables.
}

