% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{node.sizes}
\alias{node.sizes}
\alias{node.sizes,BN}
\alias{node.sizes,BN-method}
\alias{node.sizes,BNDataset}
\alias{node.sizes,BNDataset-method}
\title{get size of the variables of an object.}
\usage{
node.sizes(x)

\S4method{node.sizes}{BNDataset}(x)

\S4method{node.sizes}{BN}(x)
}
\arguments{
\item{x}{an object.}
}
\value{
vector contaning the size of each variable of the desired object.
}
\description{
Return a list containing the size of the variables of an object. It is the actual cardinality
of discrete variables, and the cardinality of the discretized variable for continuous variables.
}

