% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{num.boots<-}
\alias{num.boots<-}
\alias{num.boots<-,BNDataset-method}
\title{set number of bootstrap samples of a \code{\link{BNDataset}}.}
\usage{
num.boots(x) <- value
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}

\item{value}{the number of bootstrap samples.}
}
\description{
Set the length of the list of samples of a dataset computed using bootstrap.
}

