% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{num.boots}
\alias{num.boots}
\alias{num.boots,BNDataset}
\alias{num.boots,BNDataset-method}
\title{get number of bootstrap samples of a \code{\link{BNDataset}}.}
\usage{
num.boots(x)

\S4method{num.boots}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}
}
\value{
the number of bootstrap samples.
}
\description{
Return the number of bootstrap samples computed from a dataset.
}

