% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\docType{methods}
\name{num.items<-}
\alias{num.items<-}
\alias{num.items<-,BNDataset-method}
\title{set number of items of a \code{\link{BNDataset}}.}
\usage{
num.items(x) <- value

\S4method{num.items}{BNDataset}(x) <- value
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}

\item{value}{number of items of the desired dataset.}
}
\description{
Set the number of observed items (rows) in a dataset.
}
