% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\name{num.items}
\alias{num.items}
\alias{num.items,BNDataset-method}
\alias{num.items,BNDataset}
\title{get number of items of a \code{\link{BNDataset}}.}
\usage{
num.items(x)

\S4method{num.items}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}
}
\value{
number of items of the desired dataset.
}
\description{
Return the number of items in a dataset, that is, the number of rows in its data slot.
}
