% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R,
%   R/InferenceEngine-methods.R
\docType{methods}
\name{num.nodes<-}
\alias{num.nodes<-}
\alias{num.nodes<-,BN-method}
\alias{num.nodes<-,InferenceEngine-method}
\title{set number of nodes of an object.}
\usage{
num.nodes(x) <- value

\S4method{num.nodes}{BN}(x) <- value

\S4method{num.nodes}{InferenceEngine}(x) <- value
}
\arguments{
\item{x}{an object.}

\item{value}{the number of nodes in the object.}
}
\description{
Set the number of nodes of an object of type \code{\link{BN}} (number of nodes of the network)
or \code{\link{InferenceEngine}} (where parameter contains the number of nodes of the junction tree).
}
