% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R,
%   R/InferenceEngine-methods.R
\name{num.nodes}
\alias{num.nodes}
\alias{num.nodes,BN-method}
\alias{num.nodes,BN}
\alias{num.nodes,InferenceEngine-method}
\alias{num.nodes,InferenceEngine}
\title{get number of nodes of an object.}
\usage{
num.nodes(x)

\S4method{num.nodes}{BN}(x)

\S4method{num.nodes}{InferenceEngine}(x)
}
\arguments{
\item{x}{an object.}
}
\value{
number of nodes of the desired object.
}
\description{
Return the name of an object, of class \code{\link{BN}} or \code{\link{InferenceEngine}}.
}
