% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R,
%   R/BNDataset-methods.R
\docType{methods}
\name{num.time.steps<-}
\alias{num.time.steps<-}
\alias{num.time.steps<-,BN-method}
\alias{num.time.steps<-,BNDataset-method}
\title{set number of time steps of a \code{\link{BN}} or a \code{\link{BNDataset}}.}
\usage{
num.time.steps(x) <- value

\S4method{num.time.steps}{BN}(x) <- value

\S4method{num.time.steps}{BNDataset}(x) <- value
}
\arguments{
\item{x}{a \code{\link{BN}} or a \code{\link{BNDataset}} object.}

\item{value}{the number of time steps.}
}
\description{
Set the number of time steps of a dataset.
}
