% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R,
%   R/BNDataset-methods.R
\name{num.time.steps}
\alias{num.time.steps}
\alias{num.time.steps,BN-method}
\alias{num.time.steps,BN}
\alias{num.time.steps,BNDataset-method}
\alias{num.time.steps,BNDataset}
\title{get number of time steps observed in a \code{\link{BN}} or a \code{\link{BNDataset}}.}
\usage{
num.time.steps(x)

\S4method{num.time.steps}{BN}(x)

\S4method{num.time.steps}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BN}} or a \code{\link{BNDataset}} object.}
}
\value{
the number of time steps.
}
\description{
Return the number of time steps observed in a dataset.
}
