% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{num.variables<-}
\alias{num.variables<-}
\alias{num.variables<-,BNDataset-method}
\title{set number of variables of a \code{\link{BNDataset}}.}
\usage{
num.variables(x) <- value

\S4method{num.variables}{BNDataset}(x) <- value
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}

\item{value}{number of variables of the dataset.}
}
\description{
Set the number of variables observed in a dataset.
}

