% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/InferenceEngine-methods.R
\docType{methods}
\name{observations<-}
\alias{observations<-}
\alias{observations<-,InferenceEngine-method}
\title{set the list of observations of an \code{\link{InferenceEngine}}.}
\usage{
observations(x) <- value

\S4method{observations}{InferenceEngine}(x) <- value
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}

\item{value}{the list of observations of the \code{\link{InferenceEngine}}.}
}
\description{
Add a list of observations to an InferenceEngine, using a list of observations composed by the two following vectors:
\itemize{
\item{\code{observed.vars}}{vector of observed variables;}
\item{\code{observed.vals}}{vector of values observed for the variables in \code{observed.vars} in the corresponding position.}
}
}
\details{
Replace previous list of observations, if present. In order to add evidence, and not just replace it,
one must use the \code{\link{add.observations<-}} method.

In case of multiple observations of the same variable, the last observation is the one used, as the most recent.
}
\seealso{
\code{\link{add.observations<-}}
}
