% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BN-methods.R
\name{plot}
\alias{plot}
\alias{plot,BN}
\alias{plot.BN}
\alias{plot.BN,BN}
\title{plot a \code{\link{BN}} as a picture.}
\usage{
\method{plot}{BN}(x, ..., use.node.names = TRUE, frac = 0.2,
  max.weight = max(dag(x)), node.col = rep("white", ncol(dag(x))),
  plot.wpdag = FALSE)
}
\arguments{
\item{x}{a \code{\link{BN}} object.}

\item{...}{potential further arguments for methods.}

\item{use.node.names}{\code{TRUE} if node names have to be printed. If \code{FALSE}, numbers are used instead.}

\item{frac}{fraction}

\item{max.weight}{max.weight}

\item{node.col}{list of (\code{R}) colors for the nodes.}

\item{plot.wpdag}{if \code{TRUE} plot the network according to the WPDAG computed using bootstrap instead of the DAG.}
}
\description{
plot a \code{\link{BN}} as a picture.
}

