% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R, R/BNDataset-methods.R, R/InferenceEngine-methods.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,BN}
\alias{print,BN-method}
\alias{print,BNDataset}
\alias{print,BNDataset-method}
\alias{print,InferenceEngine}
\alias{print,InferenceEngine-method}
\title{print an object to \code{stdout}.}
\usage{
print(x, ...)

\S4method{print}{BN}(x, ...)

\S4method{print}{BNDataset}(x, show.raw.data = FALSE,
  show.imputed.data = FALSE, ...)

\S4method{print}{InferenceEngine}(x, engine = "jt", ...)
}
\arguments{
\item{x}{an object.}

\item{...}{potential other arguments.}

\item{show.raw.data}{when \code{x} is a \code{\link{BNDataset}}, print also raw dataset, if available.}

\item{show.imputed.data}{when \code{x} is a \code{\link{BNDataset}}, print also imputed dataset, if available.}

\item{engine}{when \code{x} is an \code{\link{InferenceEngine}}, specify the inference engine to be shown.
Currently only \code{engine = 'jt'} is supported.}
}
\description{
print an object to \code{stdout}.
}

