% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R,
%   R/BNDataset-methods.R
\docType{methods}
\name{quantiles<-}
\alias{quantiles<-}
\alias{quantiles<-,BN-method}
\alias{quantiles<-,BNDataset-method}
\title{set the list of quantiles of an object.}
\usage{
quantiles(x) <- value

\S4method{quantiles}{BN}(x) <- value

\S4method{quantiles}{BNDataset}(x) <- value
}
\arguments{
\item{x}{a \code{\link{BN}} or \code{\link{BNDataset}}.}

\item{value}{a list of vectors.}
}
\description{
Set the list of quantiles of a \code{\link{BN}} or a \code{\link{BNDataset}}.
}
\details{
It is used when a discretization needs to be performed.
}
\seealso{
\code{\link{quantiles}}
}
