% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R,
%   R/BNDataset-methods.R
\name{quantiles}
\alias{quantiles}
\alias{quantiles,BN-method}
\alias{quantiles,BN}
\alias{quantiles,BNDataset-method}
\alias{quantiles,BNDataset}
\title{get the list of quantiles of an object.}
\usage{
quantiles(x)

\S4method{quantiles}{BN}(x)

\S4method{quantiles}{BNDataset}(x)
}
\arguments{
\item{x}{a list of vectors.}
}
\value{
the list of quantiles of the \code{\link{BN}} of \code{\link{BNDataset}}.
}
\description{
Return the list of quantiles of a \code{\link{BN}} or a \code{\link{BNDataset}}. It is set when a discretization needs to be performed.
}
\details{
Output is a list of \code{\link{num.nodes}} vectors, one per variable. Each vector is \code{NULL}
if the corresponding variable is discrete in the original dataset, and contains the cut points for the quantiles
if the corresponding variable is continuous.
}
