% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{query}
\alias{query}
\alias{query,BN}
\alias{query,BN-method}
\title{query BN given observations}
\usage{
query(x, observed.vars = c(), observed.vals = c(), ...)

\S4method{query}{BN}(x, observed.vars, observed.vals)
}
\arguments{
\item{x}{a BN.}

\item{observed.vars}{vector of observed variables.}

\item{observed.vals}{vector of observed values for corresponding variables in \code{observed.vars}.}

\item{...}{potential further arguments for method.}
}
\value{
most probable values given observations
}
\description{
query BN given observations
}

