% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\name{raw.data}
\alias{raw.data}
\alias{raw.data,BNDataset-method}
\alias{raw.data,BNDataset}
\title{get raw data of a BNDataset.}
\usage{
raw.data(x)

\S4method{raw.data}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}
}
\description{
Return raw data contained in a \code{\link{BNDataset}} object, if any.
}
\seealso{
\code{\link{has.raw.data}}, \code{\link{has.imputed.data}}
}
