% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{read.bif}
\alias{read.bif}
\alias{read.bif,character}
\alias{read.bif,character-method}
\title{Read a network from a \code{.bif} file.}
\usage{
read.bif(x)

\S4method{read.bif}{character}(x)
}
\arguments{
\item{x}{the \code{.bif} file, with absolute/relative position.}
}
\value{
a \code{\link{BN}} object.
}
\description{
Read a network described in a \code{.bif}-formatted file, and
build a \code{\link{BN}} object.
}
\details{
The method relies on a coherent ordering of variable values and parameters in the file.
}

