% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/io-methods.R
\docType{methods}
\name{read.dataset}
\alias{read.dataset}
\alias{read.dataset,BNDataset,character,character}
\alias{read.dataset,BNDataset,character,character-method}
\title{Read a dataset from file.}
\usage{
read.dataset(object, data.file, header.file, data.with.header = FALSE,
  na.string.symbol = "?", sep.symbol = "", starts.from = 1)

\S4method{read.dataset}{BNDataset,character,character}(object, data.file,
  header.file, data.with.header = FALSE, na.string.symbol = "?",
  sep.symbol = "", starts.from = 1)
}
\arguments{
\item{object}{the \code{\link{BNDataset}} object.}

\item{data.file}{the \code{data} file.}

\item{header.file}{the \code{header} file.}

\item{data.with.header}{\code{TRUE} if the first row of \code{dataset} file is an header (e.g. it contains the variable names).}

\item{na.string.symbol}{character that denotes \code{NA} in the dataset.}

\item{sep.symbol}{separator among values in the dataset.}

\item{starts.from}{starting value for entries in the dataset (observed values, default is 1).}
}
\description{
File has to be in format (describe...)
}
\examples{
\dontrun{
dataset <- BNDataset()
dataset <- read.dataset(dataset, header="file.header", dataset="file.data")
}
}
\seealso{
BNDataset
}

