% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/io-methods.R
\name{read.dsc}
\alias{read.dsc}
\alias{read.dsc,character-method}
\alias{read.dsc,character}
\title{Read a network from a \code{.dsc} file.}
\usage{
read.dsc(x)

\S4method{read.dsc}{character}(x)
}
\arguments{
\item{x}{the \code{.dsc} file, with absolute/relative position.}
}
\value{
a \code{\link{BN}} object.
}
\description{
Read a network described in a \code{.dsc}-formatted file, and
build a \code{\link{BN}} object.
}
\details{
The method relies on a coherent ordering of variable values and parameters in the file.
}
