% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/io-methods.R
\name{read.net}
\alias{read.net}
\alias{read.net,character-method}
\alias{read.net,character}
\title{Read a network from a \code{.net} file.}
\usage{
read.net(x)

\S4method{read.net}{character}(x)
}
\arguments{
\item{x}{the \code{.net} file, with absolute/relative position.}
}
\value{
a \code{\link{BN}} object.
}
\description{
Read a network described in a \code{.net}-formatted file, and
build a \code{\link{BN}} object.
}
\details{
The method relies on a coherent ordering of variable values and parameters in the file.
}
