% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{sample.dataset}
\alias{sample.dataset}
\alias{sample.dataset,BN}
\alias{sample.dataset,BN-method}
\alias{sample.dataset,InferenceEngine}
\alias{sample.dataset,InferenceEngine-method}
\title{sample a \code{\link{BNDataset}} from a network of an inference engine.}
\usage{
sample.dataset(x, n = 100)

\S4method{sample.dataset}{BN}(x, n = 100)

\S4method{sample.dataset}{InferenceEngine}(x, n = 100)
}
\arguments{
\item{x}{a \code{\link{BN}} or \code{\link{InferenceEngine}} object.}

\item{n}{number of items to sample.}
}
\value{
a \code{\link{BNDataset}}
}
\description{
sample a \code{\link{BNDataset}} from a network of an inference engine.
}

