% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R, R/InferenceEngine-methods.R
\docType{methods}
\name{sample.row}
\alias{sample.row}
\alias{sample.row,BN}
\alias{sample.row,BN-method}
\alias{sample.row,InferenceEngine}
\alias{sample.row,InferenceEngine-method}
\title{sample a row vector of values for a network.}
\usage{
sample.row(x)

\S4method{sample.row}{BN}(x)

\S4method{sample.row}{InferenceEngine}(x)
}
\arguments{
\item{x}{a \code{\link{BN}} or \code{\link{InferenceEngine}} object.}
}
\value{
a vector of values.
}
\description{
sample a row vector of values for a network.
}

