% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{save.to.eps}
\alias{save.to.eps}
\alias{save.to.eps,BN,character}
\alias{save.to.eps,BN,character-method}
\title{save a \code{\link{BN}} picture as \code{.eps} file.}
\usage{
save.to.eps(x, filename)

\S4method{save.to.eps}{BN,character}(x, filename)
}
\arguments{
\item{x}{a \code{\link{BN}} object}

\item{filename}{name (with path, if needed) of the file to be created}
}
\description{
Save an image of a Bayesian Network as an \code{.eps} file.
}
\examples{
\dontrun{
save.to.eps(x, "out.eps")
}
}
\seealso{
\code{\link{plot}}
}

