% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{scoring.func<-}
\alias{scoring.func<-}
\alias{scoring.func<-,BN-method}
\title{Set the scoring function used to learn the structure of a network.}
\usage{
scoring.func(x) <- value

\S4method{scoring.func}{BN}(x) <- value
}
\arguments{
\item{x}{the \code{\link{BN}} object.}

\item{value}{the scoring function used.}
}
\value{
updated BN.
}
\description{
Set the scoring function used in the \code{\link{learn.structure}} method.
}

