% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{sem}
\alias{sem}
\alias{sem,BN,BNDataset-method}
\alias{sem,InferenceEngine,BNDataset}
\title{Structural Expectation-Maximization algorithm.}
\usage{
sem(x, dataset, struct.threshold = 10, param.threshold = 0.001,
  scoring.func = "BIC", alpha = 0.05, ess = 1, bootstrap = FALSE,
  layering = c(), max.fanin.layers = NULL,
  max.fanin = num.variables(dataset), cont.nodes = c(),
  use.imputed.data = FALSE, use.cpc = T, ...)

\S4method{sem}{BN,BNDataset}(x, dataset, struct.threshold = 10,
  param.threshold = 0.001, scoring.func = "BIC", alpha = 0.05, ess = 1,
  bootstrap = FALSE, layering = c(), max.fanin.layers = NULL,
  max.fanin = num.variables(dataset), cont.nodes = c(),
  use.imputed.data = FALSE, use.cpc = TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{BN}} object.}

\item{dataset}{observed dataset with missing values for the Bayesian Network of \code{x}.}

\item{struct.threshold}{threshold for convergence of the structure learning step, used as stopping criterion.}

\item{param.threshold}{threshold for convergence of the parameter learning step, used as stopping criterion.}

\item{scoring.func}{the scoring function to use. Currently, one among \code{AIC} and \code{BIC}
(default - \code{BDeu} currently not supported for \code{sem}).}

\item{alpha}{confidence threshold (only for \code{mmhc}).}

\item{ess}{Equivalent Sample Size value.}

\item{bootstrap}{\code{TRUE} to use bootstrap samples.}

\item{layering}{vector containing the layers each node belongs to (only for \code{sm}).}

\item{max.fanin.layers}{matrix of available parents in each layer (only for \code{sm}).}

\item{max.fanin}{maximum number of parents for each node (only for \code{sm}).}

\item{cont.nodes}{vector containing the index of continuous variables.}

\item{use.imputed.data}{\code{TRUE} to learn the structure from the imputed dataset
(if available, a check is performed). Default is to use raw dataset}

\item{use.cpc}{(when using \code{mmhc}) compute Candidate Parent-and-Children sets instead of
starting the Hill Climbing from an empty graph.}

\item{...}{further potential arguments for method.}
}
\value{
a list containing: an \code{\link{InferenceEngine}} with a new updated network (\code{"InferenceEngine"}),
        and the imputed dataset (\code{"BNDataset"}).
}
\description{
Learn structure and parameters of a network with the Structural EM algorithm.
}
\examples{
\dontrun{
sem(x, dataset)
}
}

