% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{shd}
\alias{shd}
\title{compute the Structural Hamming Distance between two adjacency matrices.}
\usage{
shd(g1, g2)
}
\arguments{
\item{g1}{first adjacency matrix.}

\item{g2}{second adjacency matrix.}
}
\description{
Compute the Structural Hamming Distance between two adjacency matrices, that is,
the distance, in terms of edges, between two network structures. The lower the \code{shd},
the more similar are the two network structures.
}
