% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,AllTheClasses-method}
\alias{show,BN-method}
\alias{show,BNDataset-method}
\alias{show,InferenceEngine-method}
\title{Show method for objects.}
\usage{
show(object)
}
\arguments{
\item{object}{an object.}
}
\description{
The \code{show} method allows to provide a custom aspect for the output that is generated
when the name of an instance is gives as command in an R session.
}
