% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R
\docType{methods}
\name{struct.algo<-}
\alias{struct.algo<-}
\alias{struct.algo<-,BN-method}
\title{Set the algorithm used to learn the structure of a network.}
\usage{
struct.algo(x) <- value

\S4method{struct.algo}{BN}(x) <- value
}
\arguments{
\item{x}{the \code{\link{BN}} object.}

\item{value}{the scoring function used.}
}
\value{
updated BN.
}
\description{
Set the algorithm used in the \code{\link{learn.structure}} method.
}
