% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{struct.algo}
\alias{struct.algo}
\alias{struct.algo,BN}
\alias{struct.algo,BN-method}
\title{Read the algorithm used to learn the structure of a network.}
\usage{
struct.algo(x)

\S4method{struct.algo}{BN}(x)
}
\arguments{
\item{x}{the \code{\link{BN}} object.}
}
\value{
the structure learning algorithm used.
}
\description{
Read the algorithm used in the \code{\link{learn.structure}} method.
Outcome is meaningful only if the structure of a network has been learnt.
}

