% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{tune.knn.impute}
\alias{tune.knn.impute}
\title{tune the parameter k of the knn algorithm used in imputation.}
\usage{
tune.knn.impute(
  data,
  cat.var = 1:ncol(data),
  k.min = 1,
  k.max = 20,
  frac.miss = 0.1,
  n.iter = 20,
  seed = 0
)
}
\arguments{
\item{data}{a numerical matrix.}

\item{cat.var}{vector containing the categorical variables}

\item{k.min}{minimum value for k}

\item{k.max}{maximum value for k}

\item{frac.miss}{fraction of missing values to add}

\item{n.iter}{number of iterations for each k}

\item{seed}{random seed}
}
\value{
matrix of error distributions
}
\description{
tune the parameter k of the knn algorithm used in imputation.
}
