% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{updated.bn}
\alias{updated.bn}
\alias{updated.bn,InferenceEngine}
\alias{updated.bn,InferenceEngine-method}
\title{get the updated \code{\link{BN}} object contained in an \code{\link{InferenceEngine}}.}
\usage{
updated.bn(x)

\S4method{updated.bn}{InferenceEngine}(x)
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}
}
\value{
the updated \code{\link{BN}} object contained in an \code{\link{InferenceEngine}}.
}
\description{
Return an updated network contained in an InferenceEngine.
}

