% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{updated.bn<-}
\alias{updated.bn<-}
\alias{updated.bn<-,InferenceEngine-method}
\title{set the updated \code{\link{BN}} object contained in an \code{\link{InferenceEngine}}.}
\usage{
updated.bn(x) <- value
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}

\item{value}{the updated \code{\link{BN}} object contained in an \code{\link{InferenceEngine}}.}
}
\description{
Add an updated network to an InferenceEngine.
}

