% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{variables<-}
\alias{variables<-}
\alias{variables<-,BN-method}
\alias{variables<-,BNDataset-method}
\title{set variables of an object.}
\usage{
variables(x) <- value
}
\arguments{
\item{x}{an object.}

\item{value}{vector containing the variable names of the object.
       Overwrites \code{num.nodes} slot if non-matching.}
}
\description{
Set the list of variable names in a \code{\link{BN}} or \code{\link{BNDataset}} object.
}

