% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R, R/BNDataset-methods.R
\docType{methods}
\name{variables}
\alias{variables}
\alias{variables,BN}
\alias{variables,BN-method}
\alias{variables,BNDataset}
\alias{variables,BNDataset-method}
\title{get variables of an object.}
\usage{
variables(x)

\S4method{variables}{BN}(x)

\S4method{variables}{BNDataset}(x)
}
\arguments{
\item{x}{an object.}
}
\value{
vector of the variables names of the desired object.
}
\description{
Get the list of variables (with their names) of a \code{\link{BN}} or \code{\link{BNDataset}}.
}

