% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{wpdag}
\alias{wpdag}
\alias{wpdag,BN}
\alias{wpdag,BN-method}
\title{get the WPDAG of an object.}
\usage{
wpdag(x)

\S4method{wpdag}{BN}(x)
}
\arguments{
\item{x}{an object.}
}
\value{
matrix contaning the WPDAG of the object.
}
\description{
Return the weighted partially directed acyclic graph of a network, when available (e.g. when bootstrap on dataset is performed).
}

