% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/io-methods.R
\docType{methods}
\name{write.dsc}
\alias{write.dsc}
\alias{write.dsc,BN}
\alias{write.dsc,BN-method}
\title{Write a network saving it in a \code{.dsc} file.}
\usage{
write.dsc(x, path = "./")

\S4method{write.dsc}{BN}(x, path = "./")
}
\arguments{
\item{x}{the \code{\link{BN}} object.}

\item{path}{the relative or absolute path of the directory of the created file.}
}
\description{
Write a network on disk, saving it in a \code{.dsc}-formatted file.
}

