% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-methods.R
\name{asia}
\alias{asia}
\title{load \code{Asia} dataset.}
\usage{
asia()
}
\value{
a BNDataset containing the \code{Child} dataset.
}
\description{
Wrapper for a loader for the \code{Asia} dataset, with only raw data.
}
\details{
The dataset has 10000 items, no missing data, so no imputation needs to be performed.
}
\examples{
dataset <- asia()
print(dataset)

}
\seealso{
\code{\link{asia_10000}}
}
