% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{bn}
\alias{bn}
\alias{bn,InferenceEngine}
\alias{bn,InferenceEngine-method}
\title{get the \code{\link{BN}} object contained in an \code{\link{InferenceEngine}}.}
\usage{
bn(x)

\S4method{bn}{InferenceEngine}(x)
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}
}
\value{
the \code{\link{BN}} object contained in an \code{\link{InferenceEngine}}.
}
\description{
Return a network contained in an InferenceEngine.
}

