% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/InferenceEngine-methods.R
\name{bn}
\alias{bn}
\alias{bn,InferenceEngine-method}
\alias{bn,InferenceEngine}
\title{get the \code{\link{BN}} object contained in an \code{\link{InferenceEngine}}.}
\usage{
bn(x)

\S4method{bn}{InferenceEngine}(x)
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}
}
\value{
the \code{\link{BN}} object contained in an \code{\link{InferenceEngine}}.
}
\description{
Return a network contained in an InferenceEngine.
}
