% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/InferenceEngine-methods.R
\docType{methods}
\name{bn<-}
\alias{bn<-}
\alias{bn<-,InferenceEngine-method}
\title{set the original \code{\link{BN}} object contained in an \code{\link{InferenceEngine}}.}
\usage{
bn(x) <- value

\S4method{bn}{InferenceEngine}(x) <- value
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}

\item{value}{the \code{\link{BN}} object contained in an \code{\link{InferenceEngine}}.}
}
\description{
Add an original network to an InferenceEngine.
}
