% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{BNDataset-class}
\alias{BNDataset}
\alias{BNDataset,BNDataset-class}
\alias{BNDataset-class}
\alias{initialize,BNDataset-method}
\title{BNDataset class.}
\usage{
\S4method{initialize}{BNDataset}(.Object, ...)

BNDataset(name = "", ...)
}
\arguments{
\item{.Object}{an empty BNDataset.}

\item{name}{name of the dataset.}

\item{...}{potential further arguments of methods.}
}
\value{
a BNDataset object.

BNDataset object.
}
\description{
Contains the all of the data that can be extracted from a given dataset:
raw data, imputed data, raw and imputed data with bootstrap.

initialize a \code{\link{BNDataset}} object.

constructor for \code{\link{BNDataset}} object
}
\details{
Dataset should be provided in the following format... (describe)
}
\section{Slots}{

\describe{
  \item{\code{name}:}{name of the dataset}
  \item{\code{header.file}:}{name and location of the header file}
  \item{\code{data.file}:}{name and location of the data file}
  \item{\code{variables}:}{names of the variables in the network}
  \item{\code{node.sizes}:}{cardinality of each variable of the network}
  \item{\code{num.variables}:}{number of variables (columns) in the dataset}
  \item{\code{discreteness}:}{\code{TRUE} if variable is discrete, \code{FALSE} if variable is continue}
  \item{\code{num.items}:}{number of observations (rows) in the dataset}
  \item{\code{has.rawdata}:}{TRUE if the dataset contains data read from a file}
  \item{\code{has.impdata}:}{TRUE if the dataset contains imputed data (computed from raw data)}
  \item{\code{raw.data}:}{matrix containing raw data}
  \item{\code{imputation}:}{TRUE if it dataset contains imputed data}
  \item{\code{imputed.data}:}{matrix containing imputed data}
  \item{\code{has.boots}:}{dataset has bootstrap samples}
  \item{\code{boots}:}{list of bootstrap samples}
  \item{\code{has.imp.boots}:}{dataset has imputed bootstrap samples}
  \item{\code{imp.boots}:}{list of imputed bootstrap samples}
  \item{\code{num.boots}:}{number of bootstrap samples}
}
}
\examples{
\dontrun{
dataset <- BNDataset()
dataset <- read.dataset(dataset, "file.header", "file.data")
}
}

