% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R, R/BNDataset-methods.R
\docType{class}
\name{BNDataset-class}
\alias{BNDataset}
\alias{BNDataset,BNDataset-class}
\alias{BNDataset-class}
\alias{initialize,BNDataset-method}
\title{BNDataset class.}
\usage{
\S4method{initialize}{BNDataset}(.Object)
}
\arguments{
\item{.Object}{an empty BNDataset.}

\item{data}{raw data.frame or path/name of the file containing the raw dataset (see 'Details').}

\item{discreteness}{a vector of elements in \{\code{c},\code{d}\} for continuous and discrete variables (respectively),
or path/name of the file containing header information for the dataset (discreteness, variable names, cardinality - see 'Details').}

\item{variables}{vector of variable names.}

\item{node.sizes}{vector of variable cardinalities (for discrete variables) or quantization ranges (for continuous variables).}

\item{header.file}{the \code{header} file.}

\item{data.file}{the \code{data} file.}

\item{...}{further arguments for reading a dataset from files (see documentation for \code{read.dataset}).}
}
\value{
BNDataset object.

a BNDataset object.
}
\description{
Contains the all of the data that can be extracted from a given dataset:
raw data, imputed data, raw and imputed data with bootstrap.

initialize a \code{\link{BNDataset}} object.
}
\details{
There are two ways to build a dataset, directly providing all of the informations, and using two files.
The key informations needed are:
1. the data;
2. the state of variables (discrete or continuous);
3. the names of the variables;
4. the cardinalities of the variables (if discrete), or the number of levels they have to be quantized into
(if continuous).
Names and cardinalities/leves can be guessed by looking at the data, but it is strongly advised to provide
_all_ of the informations, in order to avoid problems later on during the execution.
Data can be provided in form of data.frame or matrix. It can contain NAs. By default, NAs are indicated with '?';
to specify a different character for NAs, it is possible to provide also the \code{na.string.symbol} parameter.

It is possible to use two files, instead of providing manually all of the info.



There are two ways to build a BNDataset: using two files containing respectively header informations
and data, and manually providing the data table and the related header informations
(variable names, cardinality and discreteness).
The two ways cannot be mixed up. If the method of choice is the file-based one, both the files
must be provided. Conversely, all of the \code{data}, \code{variables}, \code{node.sizes}
and \code{discreteness} parameters has to be given, and the files parameters must be ignored.
In any case, the names of the parameters used must be provided to the constructor.

bnstruct requires the data files to be in a format subsequently described.
The actual data has to be in (a text file containing data in) tabular format, one tuple per row,
with the values for each variable separated by a space or a tab. Values for each variable have to be
numbers, starting from \code{1} in case of discrete variables.
Data files can have a first row containing the names of the corresponding variables.

In addition to the data file, a header file containing additional informations can also be provided.
An header file has to be composed by three rows of tab-delimited values:
1. list of names of the variables, in the same order of the data file;
2. a list of integers representing the cardinality of the variables, in case of discrete variables,
  or the number of levels each variable has to be quantized in, in case of continuous variables;
3. a list that indicates, for each variable, if the variable is continuous
  (\code{c} or \code{C}), and thus has to be quantized before learning,
  or discrete (\code{d} or \code{D}).

In case of need of more advanced options when reading a dataset from files, please refer to the
documentation of the \code{\link{read.dataset}} method. Imputation and bootstrap are also available
as separate routines (\code{\link{impute}} and \code{\link{bootstrap}}, respectively).
}
\section{Slots}{

\describe{
  \item{\code{name}:}{name of the dataset}
  \item{\code{header.file}:}{name and location of the header file}
  \item{\code{data.file}:}{name and location of the data file}
  \item{\code{variables}:}{names of the variables in the network}
  \item{\code{node.sizes}:}{cardinality of each variable of the network}
  \item{\code{num.variables}:}{number of variables (columns) in the dataset}
  \item{\code{discreteness}:}{\code{TRUE} if variable is discrete, \code{FALSE} if variable is continue}
  \item{\code{num.items}:}{number of observations (rows) in the dataset}
  \item{\code{has.rawdata}:}{TRUE if the dataset contains data read from a file}
  \item{\code{has.impdata}:}{TRUE if the dataset contains imputed data (computed from raw data)}
  \item{\code{raw.data}:}{matrix containing raw data}
  \item{\code{imputation}:}{TRUE if it dataset contains imputed data}
  \item{\code{imputed.data}:}{matrix containing imputed data}
  \item{\code{has.boots}:}{dataset has bootstrap samples}
  \item{\code{boots}:}{list of bootstrap samples}
  \item{\code{has.imp.boots}:}{dataset has imputed bootstrap samples}
  \item{\code{imp.boots}:}{list of imputed bootstrap samples}
  \item{\code{num.boots}:}{number of bootstrap samples}
}
}
\examples{
\dontrun{
# create from files
dataset <- read.dataset("file.header", "file.data")

# other way: create from raw dataset and metadata
data <- matrix(c(1:16), nrow = 4, ncol = 4)
dataset <- BNDataset(name = "MyData", data = data,
                     variables = c("a", "b", "c", "d"),
                     node.sizes = c(4,8,12,16),
                     discreteness = rep('d',4))
}
}
\seealso{
read.dataset, impute, bootstrap
}

