% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{BNParams-class}
\alias{BNParams}
\alias{BNParams,BNParams-class}
\alias{BNParams-class}
\alias{initialize,BNParams-method}
\title{BNParams class}
\usage{
\S4method{initialize}{BN}(.Object, ...)

BNParams(args, ...)
}
\arguments{
\item{.Object}{a BNParams}

\item{...}{potential further arguments of methods.}

\item{args}{the list of command line parameters passed to the R script.}
}
\value{
BNParams object.
}
\description{
Contains all of the parameters available for the algorithms contained in the package.

Takes the list of command line parameters passed in input, and processes it.
Does not check for their correctness (it's assumed they are correct).
}
\section{Slots}{

\describe{
\item{\code{ess}}{Equivalent Sample Size value.}

\item{\code{alpha}}{confidence threshold (for \code{mmhc} and \code{eocp}).}

\item{\code{learning.algo}}{algorithm to use for structure learning: one among \code{mmhc} (Max-min Hill Climbing, default),
\code{sm} (Silander-Myllymaki) and \code{eocp} (Edge Orientation with candidate parents).}

\item{\code{scoring.func}}{scoring function to be used for structure learning: one among \code{BDeu} (default),
\code{AIC} and \code{BIC}.}

\item{\code{max.parents}}{maximum number of parents for each node in the network (a value <0 means "all of the other nodes of the network").}

\item{\code{num.boots}}{number of bootstrap samples to generate, if needed.}

\item{\code{k.impute}}{number of neighbours to be used; for discrete variables we use mode, for continuous variables the median value is instead taken.}

\item{\code{seed}}{random seed.}

\item{\code{em_convergence}}{threshold for convergence of the EM algorithm as difference in the CPTs between two successive iterations,
used as stopping criterion.}

\item{\code{sem_convergence}}{threshold for convergence of the SEM algorithm as Structural Hamming Distance between the networks
computed in two successive iterations, used as stopping criterion.}

\item{\code{CPX_PARAM_SCRIND}}{"numeric",}

\item{\code{CPX_PARAM_PRELINEAR}}{"numeric",}

\item{\code{CPX_PARAM_PREIND}}{"numeric",}

\item{\code{CPX_PARAM_MIPCBREDLP}}{"numeric",}

\item{\code{CPX_PARAM_MIPSEARCH}}{"numeric",}

\item{\code{cplex_presolve_algo}}{"numeric"}
}}
\examples{
\dontrun{
params <- BNParams()
}
}

