% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\docType{methods}
\name{boots<-}
\alias{boots<-}
\alias{boots<-,BNDataset-method}
\title{set list of bootstrap samples of a \code{\link{BNDataset}}.}
\usage{
boots(x) <- value

\S4method{boots}{BNDataset}(x) <- value
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}

\item{value}{the list of bootstrap samples.}
}
\description{
Add to a dataset a list of samples from raw data computed using bootstrap.
}
