% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\name{boots}
\alias{boots}
\alias{boots,BNDataset-method}
\alias{boots,BNDataset}
\title{get list of bootstrap samples of a \code{\link{BNDataset}}.}
\usage{
boots(x)

\S4method{boots}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}
}
\value{
the list of bootstrap samples.
}
\description{
Return the list of samples computed from raw data of a dataset.
}
\seealso{
\code{\link{has.boots}}, \code{\link{has.imputed.boots}}, \code{\link{imp.boots}}
}
