% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap,BNDataset}
\alias{bootstrap,BNDataset-method}
\title{Perform bootstrap.}
\usage{
bootstrap(object, num.boots = params@num.boots, seed = params@seed,
  imputation = FALSE, k.impute = params@k.impute, na.string.symbol = "?",
  ..., params)

\S4method{bootstrap}{BNDataset}(object, num.boots = params@num.boots,
  seed = params@seed, imputation = FALSE, k.impute = params@k.impute,
  na.string.symbol = "?", ..., params)
}
\arguments{
\item{object}{the \code{\link{BNDataset}} object.}

\item{num.boots}{number of sampled datasets for bootstrap.}

\item{seed}{random seed.}

\item{imputation}{\code{TRUE} if imputation has to be performed.}

\item{k.impute}{number of neighbours to be used; for discrete variables we use mode, for continuous variables the median value is instead taken (useful only if imputation == TRUE).}

\item{na.string.symbol}{character that denotes NA in the dataset (useful only if imputation == TRUE).}

\item{...}{potential further arguments of methods.}

\item{params}{a \code{\link{BNParams}} object.}
}
\description{
Create a list of \code{num.boots} samples of the original dataset.
}
\examples{
\dontrun{
dataset <- BNDataset()
dataset <- read.dataset(dataset, "file.header", "file.data", params = myparams)
dataset <- bootstrap(dataset, num.boots = 1000, params = myparams)
}
}

