% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap,BNDataset}
\alias{bootstrap,BNDataset-method}
\title{Perform bootstrap.}
\usage{
bootstrap(object, ...)

\S4method{bootstrap}{BNDataset}(object, num.boots = 100, seed = 0,
  imputation = FALSE, k.impute = 10, na.string.symbol = "?", ...)
}
\arguments{
\item{object}{the \code{\link{BNDataset}} object.}

\item{num.boots}{number of sampled datasets for bootstrap.}

\item{seed}{random seed.}

\item{imputation}{\code{TRUE} if imputation has to be performed.}

\item{na.string.symbol}{character that denotes NA in the dataset (useful only if imputation == TRUE).}

\item{k.impute}{number of neighbours to be used; for discrete variables we use mode, for continuous variables the median value is instead taken (useful only if imputation == TRUE).}

\item{...}{potential further arguments of methods.}
}
\description{
Create a list of \code{num.boots} samples of the original dataset.
}
\examples{
\dontrun{
dataset <- BNDataset()
dataset <- read.dataset(dataset, "file.header", "file.data")
dataset <- bootstrap(dataset, num.boots = 1000)
}
}

