% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dag}
\alias{dag}
\alias{dag,BN}
\alias{dag,BN-method}
\title{get adjacency matrix of a network.}
\usage{
dag(x)

\S4method{dag}{BN}(x)
}
\arguments{
\item{x}{an object.}
}
\value{
matrix containing the adjacency matrix of the directed acyclic graph representing
        the structure of the object.
}
\description{
Return the adjacency matrix of the directed acyclic graph representing the structure of a network.
}

