% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R
\name{dag}
\alias{dag}
\alias{dag,BN-method}
\alias{dag,BN}
\title{get adjacency matrix of a network.}
\usage{
dag(x)

\S4method{dag}{BN}(x)
}
\arguments{
\item{x}{an object.}
}
\value{
matrix containing the adjacency matrix of the directed acyclic graph representing
        the structure of the object.
}
\description{
Return the adjacency matrix of the directed acyclic graph representing the structure of a network.
}
