% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R,
%   R/BNDataset-methods.R
\docType{methods}
\name{discreteness<-}
\alias{discreteness<-}
\alias{discreteness<-,BN-method}
\alias{discreteness<-,BNDataset-method}
\title{set status (discrete or continuous) of the variables of an object.}
\usage{
discreteness(x) <- value

\S4method{discreteness}{BN}(x) <- value

\S4method{discreteness}{BNDataset}(x) <- value
}
\arguments{
\item{x}{an object.}

\item{value}{a vector of elements in \{\code{c},\code{d}\} for continuous and discrete variables (respectively).}
}
\description{
Set the list of variable status for the variables in a network or a dataset.
}
