% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{em}
\alias{em}
\alias{em,InferenceEngine,BNDataset}
\alias{em,InferenceEngine,BNDataset-method}
\title{expectation-maximization algorithm.}
\usage{
em(x, dataset, ...)

\S4method{em}{InferenceEngine,BNDataset}(x, dataset, threshold = 0.001,
  k.impute = 10, ...)
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}

\item{dataset}{observed dataset with missing values for the Bayesian Network of \code{x}.}

\item{threshold}{threshold for convergence, used as stopping criterion.}

\item{k.impute}{radius of imputation.}

\item{...}{further potential arguments for method.}
}
\value{
an \code{\link{InferenceEngine}} with a new updated network.
}
\description{
Learn parameters of a network using thr Expectation-Maximization algorithm.
}
\examples{
\dontrun{
em(x, dataset)
}
}

