% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{em}
\alias{em}
\alias{em,InferenceEngine,BNDataset}
\alias{em,InferenceEngine,BNDataset-method}
\title{expectation-maximization algorithm.}
\usage{
em(x, dataset, threshold = 0.001, k.impute = 10, ...)

\S4method{em}{InferenceEngine,BNDataset}(x, dataset, threshold = 0.001,
  k.impute = 10, ...)
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}

\item{dataset}{observed dataset with missing values for the Bayesian Network of \code{x}.}

\item{threshold}{threshold for convergence, used as stopping criterion.}

\item{k.impute}{number of neighbours to be used; for discrete variables we use mode, for continuous variables the median value is instead taken.}

\item{...}{further potential arguments for method.}
}
\value{
a list containing: an \code{\link{InferenceEngine}} with a new updated network (\code{"InferenceEngine"}),
        and the imputed dataset (\code{"BNDataset"}).
}
\description{
Learn parameters of a network using the Expectation-Maximization algorithm.
}
\examples{
\dontrun{
em(x, dataset)
}
}

