% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\docType{methods}
\name{has.imputed.data}
\alias{has.imputed.data}
\alias{has.imputed.data,BNDataset}
\alias{has.imputed.data,BNDataset-method}
\title{check if a BNDataset contains impited data.}
\usage{
has.imputed.data(x)

has.imputed.data(x)

\S4method{has.imputed.data}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}

\item{x}{a \code{\link{BNDataset}} object.}
}
\value{
\code{TRUE} if dataset has imputed data.
}
\description{
Check whether a \code{\link{BNDataset}} object actually contains imputed data.

Return \code{TRUE} if the given dataset contains imputed data, \code{FALSE} otherwise.
}
\examples{
\dontrun{
x <- BNDataset()
has.imputed.data(x) # FALSE

x <- read.dataset(x, "file.header", "file.data")
has.imputed.data(x) # FALSE, since read.dataset() actually reads raw data.

x <- impute(x)
has.imputed.data(x) # TRUE
}
}
\seealso{
\code{\link{has.raw.data}}, \code{\link{raw.data}}, \code{\link{imputed.data}}

\code{\link{has.raw.data}}, \code{\link{raw.data}}, \code{\link{imputed.data}}
}

