% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{has.raw.data}
\alias{has.raw.data}
\alias{has.raw.data,BNDataset}
\alias{has.raw.data,BNDataset-method}
\title{check if a BNDataset contains raw data.}
\usage{
has.raw.data(x)

\S4method{has.raw.data}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}
}
\description{
Check whether a \code{\link{BNDataset}} object actually contains raw data.
}
\examples{
\dontrun{
x <- BNDataset()
has.raw.data(x) # FALSE

x <- read.dataset(x, "file.header", "file.data")
has.raw.data(x) # TRUE, since read.dataset() actually reads raw data.
}
}
\seealso{
\code{\link{has.data}}, \code{\link{has.imputed.data}}, \code{\link{raw.data}}, \code{\link{imputed.data}}
}

