% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\name{imp.boots}
\alias{imp.boots}
\alias{imp.boots,BNDataset-method}
\alias{imp.boots,BNDataset}
\title{get list of bootstrap samples from imputed data of a \code{\link{BNDataset}}.}
\usage{
imp.boots(x)

\S4method{imp.boots}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}
}
\value{
the list of bootstrap samples from imputed data.
}
\description{
Return the list of samples computed from raw data of a dataset.
}
\seealso{
\code{\link{has.boots}}, \code{\link{has.imputed.boots}}, \code{\link{boots}}
}
