% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{impute}
\alias{impute}
\alias{impute,BNDataset}
\alias{impute,BNDataset-method}
\title{Impute a \code{\link{BNDataset}} raw data with missing values.}
\usage{
impute(object, ...)

\S4method{impute}{BNDataset}(object, k.impute = 10)
}
\arguments{
\item{object}{the \code{\link{BNDataset}} object.}

\item{k.impute}{number of neighbours to be used; for discrete variables we use mode, for continuous variables the median value is instead taken.}

\item{...}{potential further arguments of methods.}
}
\description{
Impute a \code{\link{BNDataset}} raw data with missing values.
}
\examples{
\dontrun{
dataset <- BNDataset()
dataset <- read.dataset(dataset, "file.header", "file.data")
dataset <- impute(dataset)
}
}

