% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\name{impute}
\alias{impute}
\alias{impute,BNDataset-method}
\alias{impute,BNDataset}
\title{Impute a \code{\link{BNDataset}} raw data with missing values.}
\usage{
impute(object, k.impute = 10)

\S4method{impute}{BNDataset}(object, k.impute = 10)
}
\arguments{
\item{object}{the \code{\link{BNDataset}} object.}

\item{k.impute}{number of neighbours to be used; for discrete variables we use mode,
for continuous variables the median value is instead taken.}
}
\description{
Impute a \code{\link{BNDataset}} raw data with missing values.
}
\examples{
\dontrun{
dataset <- BNDataset("file.data", "file.header")
dataset <- impute(dataset)
}

}
