% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\docType{methods}
\name{imputed.data<-}
\alias{imputed.data<-}
\alias{imputed.data<-,BNDataset-method}
\title{add imputed data.}
\usage{
imputed.data(x) <- value

\S4method{imputed.data}{BNDataset}(x) <- value
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}

\item{value}{a matrix of integers containing a dataset.}
}
\description{
Insert imputed data in a \code{\link{BNDataset}} object.
}
\seealso{
\code{\link{has.imputed.data}}, \code{\link{imputed.data}}, \code{\link{read.dataset}}
}
