% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/InferenceEngine-methods.R
\name{jt.cliques}
\alias{jt.cliques}
\alias{jt.cliques,InferenceEngine-method}
\alias{jt.cliques,InferenceEngine}
\title{get the list of cliques of the junction tree of an \code{\link{InferenceEngine}}.}
\usage{
jt.cliques(x)

\S4method{jt.cliques}{InferenceEngine}(x)
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}
}
\value{
the list of cliques of the junction tree contained in the \code{\link{InferenceEngine}}.
}
\description{
Return the list of cliques containing the variables associated to each node of a junction tree.
}
