% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/InferenceEngine-methods.R
\docType{methods}
\name{junction.tree<-}
\alias{junction.tree<-}
\alias{junction.tree<-,InferenceEngine-method}
\title{set the junction tree of an \code{\link{InferenceEngine}}.}
\usage{
junction.tree(x) <- value

\S4method{junction.tree}{InferenceEngine}(x) <- value
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}

\item{value}{the junction tree to be inserted in the \code{\link{InferenceEngine}}.}
}
\description{
Set the adjacency matrix of the junction tree computed for a network.
}
