% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R
\name{layering}
\alias{layering}
\alias{layering,BN-method}
\alias{layering,BN}
\title{return the layering of the nodes.}
\usage{
layering(x)

\S4method{layering}{BN}(x)
}
\arguments{
\item{x}{a \code{\link{BN}} object.}
}
\value{
a vector containing layers the nodes can be divided into.
}
\description{
Compute the topological ordering of the nodes of a network, in order to divide the network in layers.
}
\examples{
\dontrun{
dataset <- BNDataset("file.header", "file.data")
x <- BN(dataset)
x <- learn.network(x, dataset)
layering(x)
}

}
