% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R,
%   R/BNDataset-methods.R
\docType{methods}
\name{name<-}
\alias{name<-}
\alias{name<-,BN-method}
\alias{name<-,BNDataset-method}
\title{set name of an object.}
\usage{
name(x) <- value

\S4method{name}{BN}(x) <- value

\S4method{name}{BNDataset}(x) <- value
}
\arguments{
\item{x}{an object.}

\item{value}{the new name of the object.}
}
\description{
Set the \code{name} slot of an object of type \code{\link{BN}} or \code{\link{BNDataset}}.
}
