% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{name}
\alias{name}
\alias{name,BN}
\alias{name,BN-method}
\alias{name,BNDataset}
\alias{name,BNDataset-method}
\title{get name of an object.}
\usage{
name(x)

\S4method{name}{BNDataset}(x)

\S4method{name}{BN}(x)
}
\arguments{
\item{x}{an object.}
}
\value{
name of the object.
}
\description{
Return the name of an object, of class \code{\link{BN}} or \code{\link{BNDataset}}.
}

