% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R,
%   R/BNDataset-methods.R
\name{node.sizes}
\alias{node.sizes}
\alias{node.sizes,BN-method}
\alias{node.sizes,BN}
\alias{node.sizes,BNDataset-method}
\alias{node.sizes,BNDataset}
\title{get size of the variables of an object.}
\usage{
node.sizes(x)

\S4method{node.sizes}{BN}(x)

\S4method{node.sizes}{BNDataset}(x)
}
\arguments{
\item{x}{an object.}
}
\value{
vector contaning the size of each variable of the desired object.
}
\description{
Return a list containing the size of the variables of an object. It is the actual cardinality
of discrete variables, and the cardinality of the discretized variable for continuous variables.
}
