% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\docType{methods}
\name{num.boots<-}
\alias{num.boots<-}
\alias{num.boots<-,BNDataset-method}
\title{set number of bootstrap samples of a \code{\link{BNDataset}}.}
\usage{
num.boots(x) <- value

\S4method{num.boots}{BNDataset}(x) <- value
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}

\item{value}{the number of bootstrap samples.}
}
\description{
Set the length of the list of samples of a dataset computed using bootstrap.
}
