% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\name{num.variables}
\alias{num.variables}
\alias{num.variables,BNDataset-method}
\alias{num.variables,BNDataset}
\title{get number of variables of a \code{\link{BNDataset}}.}
\usage{
num.variables(x)

\S4method{num.variables}{BNDataset}(x)

\S4method{num.variables}{BNDataset}(x)
}
\arguments{
\item{x}{a \code{\link{BNDataset}} object.}
}
\value{
number of variables of the desired dataset.
}
\description{
Return the number of the variables contained in a dataset. This value corresponds to the value
of \code{\link{num.nodes}} of a network built upon the same dataset.
}
\seealso{
\code{\link{num.nodes}}
}
