% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/InferenceEngine-methods.R
\docType{methods}
\name{observations}
\alias{observations}
\alias{observations,InferenceEngine}
\alias{observations,InferenceEngine-method}
\title{get the list of observations of an \code{\link{InferenceEngine}}.}
\usage{
observations(x)

\S4method{observations}{InferenceEngine}(x)
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}
}
\value{
the list of observations of the \code{\link{InferenceEngine}}.
}
\description{
Return the list of observations added to an InferenceEngine.
}
\details{
Output is a list in the following format:
\itemize{
\item{\code{observed.vars}}{vector of observed variables;}
\item{\code{observed.vals}}{vector of values observed for the variables in \code{observed.vars} in the corresponding position.}
}
}

