% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,BN}
\alias{plot,BN-method}
\title{plot a \code{\link{BN}} as a picture.}
\usage{
plot(x, ...)

\S4method{plot}{BN}(x, use.node.names = TRUE, frac = 0.2,
  max.weight = max(dag(x)), node.col = rep("white", ncol(dag(x))),
  plot.wpdag = FALSE)
}
\arguments{
\item{x}{a \code{\link{BN}} object.}

\item{use.node.names}{\code{TRUE} if node names have to be printed. If \code{FALSE}, number are used instead.}

\item{frac}{fraction}

\item{max.weight}{max.weight}

\item{node.col}{list of (\code{R}) colors for the nodes.}

\item{plot.wpdag}{if \code{TRUE} plot the network according to the WPDAG computed using bootstrap instead of the DAG.}

\item{...}{potential further arguments of methods.}
}
\description{
Plot the network as a picture to default output.
}
\examples{
\dontrun{
plot(x, use.node.names=TRUE, frac=0.2, max.weight=1,
     node.col=c("cyan"), plot.wpdag=FALSE)
}
}

