% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN-methods.R, R/BNDataset-methods.R,
%   R/InferenceEngine-methods.R
\name{print}
\alias{print}
\alias{print.BN}
\alias{print,BN}
\alias{print.BN,BN}
\alias{print.BNDataset}
\alias{print,BNDataset}
\alias{print.BNDataset,BNDataset}
\alias{print.InferenceEngine}
\alias{print,InferenceEngine}
\alias{print.InferenceEngine,InferenceEngine}
\title{print a \code{\link{BN}}, \code{\link{BNDataset}} or \code{\link{InferenceEngine}} to \code{stdout}.}
\usage{
\method{print}{BN}(x, ...)

\method{print}{BNDataset}(x, show.raw.data = FALSE, show.imputed.data = FALSE, ...)

\method{print}{InferenceEngine}(x, engine = "jt", ...)
}
\arguments{
\item{x}{a \code{\link{BN}}, \code{\link{BNDataset}} or \code{\link{InferenceEngine}}.}

\item{...}{potential other arguments.}

\item{show.raw.data}{if \code{x} is a \code{\link{BNDataset}}, print also raw dataset, if available.}

\item{show.imputed.data}{if \code{x} is a \code{\link{BNDataset}}, print also imputed dataset, if available.}

\item{engine}{if \code{x} is an \code{\link{InferenceEngine}}, specify the inference engine to be shown. Currently only \code{engine = 'jt'} is supported.}
}
\description{
print a \code{\link{BN}}, \code{\link{BNDataset}} or \code{\link{InferenceEngine}} to \code{stdout}.
}
