% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{raw.data<-}
\alias{raw.data<-}
\alias{raw.data<-,BNDataset-method}
\title{add raw data.}
\usage{
raw.data(x) <- value

\S4method{raw.data}{BNDataset}(x) <- value
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}

\item{value}{a matrix of integers containing a dataset.}
}
\description{
Insert raw data in a \code{\link{BNDataset}} object.
}
\details{
Users are encouraged to not use this method whenever possible, in favour of \code{\link{read.dataset}}.
}
\seealso{
\code{\link{has.data}}, \code{\link{has.raw.data}}, \code{\link{get.data}}, \code{\link{read.dataset}}
}

