% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BNDataset-methods.R
\docType{methods}
\name{raw.data<-}
\alias{raw.data<-}
\alias{raw.data<-,BNDataset-method}
\title{add raw data.}
\usage{
raw.data(x) <- value

\S4method{raw.data}{BNDataset}(x) <- value
}
\arguments{
\item{x}{a \code{\link{BNDataset}}.}

\item{value}{a matrix of integers containing a dataset.}
}
\description{
Insert raw data in a \code{\link{BNDataset}} object.
}
\seealso{
\code{\link{has.raw.data}}, \code{\link{raw.data}}, \code{\link{read.dataset}}
}
