% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/io-methods.R
\name{read.bif}
\alias{read.bif}
\alias{read.bif,character-method}
\alias{read.bif,character}
\title{Read a network from a \code{.bif} file.}
\usage{
read.bif(x)

\S4method{read.bif}{character}(x)
}
\arguments{
\item{x}{the \code{.bif} file, with absolute/relative position.}
}
\value{
a \code{\link{BN}} object.
}
\description{
Read a network described in a \code{.bif}-formatted file, and
build a \code{\link{BN}} object.
}
\details{
The method relies on a coherent ordering of variable values and parameters in the file.
}
