% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{read.dataset}
\alias{read.dataset}
\alias{read.dataset,BNDataset,character,character}
\alias{read.dataset,BNDataset,character,character-method}
\title{Read a dataset from file.}
\usage{
read.dataset(object, header.file, data.file, imputation = FALSE,
  header.flag = FALSE, na.string.symbol = "?", sep.symbol = "",
  k.impute = params@k.impute, bootstrap = FALSE,
  num.boots = params@num.boots, seed = params@seed, starts.from = 0, ...,
  params)

\S4method{read.dataset}{BNDataset,character,character}(object, header.file,
  data.file, imputation = FALSE, header.flag = FALSE,
  na.string.symbol = "?", sep.symbol = "", k.impute = params@k.impute,
  bootstrap = FALSE, num.boots = params@num.boots, seed = params@seed,
  starts.from = 0, ..., params)
}
\arguments{
\item{object}{the \code{\link{BNDataset}} object.}

\item{header.file}{the \code{header} file.}

\item{data.file}{the \code{data} file.}

\item{imputation}{\code{TRUE} if imputation has to be performed.}

\item{header.flag}{\code{TRUE} if the first row of \code{dataset} file is an header (e.g. it contains the variable names).}

\item{na.string.symbol}{character that denotes \code{NA} in the dataset.}

\item{sep.symbol}{separator among values in the dataset.}

\item{k.impute}{number of neighbours to be used; for discrete variables we use mode, for continuous variables the median value is instead taken (useful only if imputation == TRUE).}

\item{bootstrap}{\code{TRUE} if bootstrap has to be performed; prepares a list of datasets sampled from the original one.}

\item{num.boots}{number of sampled datasets for bootstrap (useful only if bootstrap == TRUE).}

\item{seed}{random seed (useful only if bootstrap == TRUE).}

\item{starts.from}{starting value for entries in the dataset (observed values, default is 0).}

\item{...}{potential further arguments of methods.}

\item{params}{a \code{\link{BNParams}} object.}
}
\description{
File has to be in format (describe...)
}
\examples{
\dontrun{
dataset <- BNDataset()
dataset <- read.dataset(dataset, header="file.header", dataset="file.data", params = myparams)
}
}

