% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{read.dataset}
\alias{read.dataset}
\alias{read.dataset,BNDataset,character,character}
\alias{read.dataset,BNDataset,character,character-method}
\title{Read a dataset from file.}
\usage{
read.dataset(object, header.file, data.file, ...)

\S4method{read.dataset}{BNDataset,character,character}(object, header.file,
  data.file, imputation = FALSE, header.flag = FALSE,
  na.string.symbol = "?", sep.symbol = "", k.impute = 10,
  bootstrap = FALSE, num.boots = 100, seed = 0, starts.from = 0, ...)
}
\arguments{
\item{object}{the \code{\link{BNDataset}} object.}

\item{header.file}{the \code{header} file.}

\item{data.file}{the \code{data} file.}

\item{na.string.symbol}{character that denotes \code{NA} in the dataset.}

\item{sep.symbol}{separator among values in the dataset.}

\item{header.flag}{\code{TRUE} if the first row of \code{dataset} file is an header (e.g. it contains the variable names).}

\item{imputation}{\code{TRUE} if imputation has to be performed.}

\item{k.impute}{number of neighbours to be used; for discrete variables we use mode, for continuous variables the median value is instead taken (useful only if imputation == TRUE).}

\item{bootstrap}{\code{TRUE} if bootstrap has to be performed; prepares a list of datasets sampled from the original one.}

\item{num.boots}{number of sampled datasets for bootstrap (useful only if bootstrap == TRUE).}

\item{seed}{random seed (useful only if bootstrap == TRUE).}

\item{starts.from}{starting value for entries in the dataset (observed values, default is 0).}

\item{...}{potential further arguments of methods.}
}
\description{
File has to be in format (describe...)
}
\examples{
\dontrun{
dataset <- BNDataset()
dataset <- read.dataset(dataset, header="file.header", dataset="file.data")
}
}

