% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R
\name{scoring.func}
\alias{scoring.func}
\alias{scoring.func,BN-method}
\alias{scoring.func,BN}
\title{Read the scoring function used to learn the structure of a network.}
\usage{
scoring.func(x)

\S4method{scoring.func}{BN}(x)
}
\arguments{
\item{x}{the \code{\link{BN}} object.}
}
\value{
the scoring function used.
}
\description{
Read the scoring function used in the \code{\link{learn.structure}} method.
Outcome is meaningful only if the structure of a network has been learnt.
}
