% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{sem}
\alias{sem}
\alias{sem,InferenceEngine,BNDataset}
\alias{sem,InferenceEngine,BNDataset-method}
\title{Structural Expectation-Maximization algorithm.}
\usage{
sem(x, dataset, struct.threshold = 10, param.threshold = 0.001,
  k.impute = 10, algo = "mmhc", scoring.func = "BDeu", alpha = 0.05,
  ess = 1, bootstrap = FALSE, layering = c(), max.fanin.layers = NULL,
  max.fanin = num.variables(dataset), cont.nodes = c(), raw.data = FALSE,
  num.boots = 100, imputation = TRUE, na.string.symbol = "?", seed = 0,
  ...)

\S4method{sem}{InferenceEngine,BNDataset}(x, dataset, struct.threshold = 10,
  param.threshold = 0.001, k.impute = 10, algo = "mmhc",
  scoring.func = "BIC", alpha = 0.05, ess = 1, bootstrap = FALSE,
  layering = c(), max.fanin.layers = NULL,
  max.fanin = num.variables(dataset), cont.nodes = c(), raw.data = FALSE,
  num.boots = 100, imputation = TRUE, na.string.symbol = "?", seed = 0,
  ...)
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}}

\item{dataset}{observed dataset with missing values for the Bayesian Network of \code{x}.}

\item{struct.threshold}{threshold for convergence of the structure learning step, used as stopping criterion.}

\item{param.threshold}{threshold for convergence of the parameter learning step, used as stopping criterion.}

\item{k.impute}{number of neighbours to be used; for discrete variables we use mode, for continuous variables the median value is instead taken.}

\item{algo}{the algorithm to use. Currently, one among \code{sm} (Silander-Myllymaki) and \code{mmhc} (Max-Min Hill Climbing, default).}

\item{scoring.func}{the scoring function to use. Currently, one among \code{BDeu}, \code{AIC}, \code{BIC}.}

\item{alpha}{confidence threshold (only for \code{mmhc}).}

\item{ess}{Equivalent Sample Size value.}

\item{bootstrap}{\code{TRUE} to use bootstrap samples.}

\item{layering}{vector containing the layers each node belongs to (only for \code{sm}).}

\item{max.fanin.layers}{matrix of available parents in each layer (only for \code{sm}).}

\item{max.fanin}{maximum number of parents for each node (only for \code{sm}).}

\item{cont.nodes}{vector containing the index of continuous variables.}

\item{raw.data}{\code{TRUE} to learn the structure from the raw dataset. Default is to use imputed dataset
(if available, otherwise the raw dataset will be used anyway).}

\item{num.boots}{number of bootstrap samples to generate, if needed.}

\item{imputation}{\code{TRUE} if imputation is needed; if \code{bootstrap=TRUE}, imputed samples will be also used.}

\item{na.string.symbol}{symbol for \code{NA} values (missing data).}

\item{seed}{random seed.}

\item{...}{further potential arguments for method.}
}
\value{
a list containing: an \code{\link{InferenceEngine}} with a new updated network (\code{"InferenceEngine"}),
        and the imputed dataset (\code{"BNDataset"}).
}
\description{
Learn struvture and parameters of a network with the Structural EM algorithm.
}
\examples{
\dontrun{
sem(x, dataset)
}
}

