% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/InferenceEngine-methods.R
\docType{methods}
\name{updated.bn<-}
\alias{updated.bn<-}
\alias{updated.bn<-,InferenceEngine-method}
\title{set the updated \code{\link{BN}} object contained in an \code{\link{InferenceEngine}}.}
\usage{
updated.bn(x) <- value

\S4method{updated.bn}{InferenceEngine}(x) <- value
}
\arguments{
\item{x}{an \code{\link{InferenceEngine}}.}

\item{value}{the updated \code{\link{BN}} object contained in an \code{\link{InferenceEngine}}.}
}
\description{
Add an updated network to an InferenceEngine.
}
