% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R,
%   R/BNDataset-methods.R
\docType{methods}
\name{variables<-}
\alias{variables<-}
\alias{variables<-,BN-method}
\alias{variables<-,BNDataset-method}
\title{set variables of an object.}
\usage{
variables(x) <- value

\S4method{variables}{BN}(x) <- value

\S4method{variables}{BNDataset}(x) <- value
}
\arguments{
\item{x}{an object.}

\item{value}{vector containing the variable names of the object.
Overwrites \code{num.nodes} slot if non-matching.}
}
\description{
Set the list of variable names in a \code{\link{BN}} or \code{\link{BNDataset}} object.
}
