% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BN-methods.R
\docType{methods}
\name{wpdag<-}
\alias{wpdag<-}
\alias{wpdag<-,BN-method}
\title{set WPDAG of the object.}
\usage{
wpdag(x) <- value

\S4method{wpdag}{BN}(x) <- value
}
\arguments{
\item{x}{an object.}

\item{value}{matrix contaning the WPDAG of the object.}
}
\description{
Set the weighted partially directed acyclic graph of a network (e.g. in case bootstrap on dataset is performed).
}
